+++ example/gpu/mm/mmTile.cu
@@ -132,7 +132,7 @@ void multiplyMM(uint32_t paramc, uint64_t * paramv, uint32_t depc, artsEdtDep_t 
     dim3 threads(SMTILE, SMTILE);
     dim3 grid((tileSize+SMTILE-1)/SMTILE, (tileSize+SMTILE-1)/SMTILE);
     
-    uint64_t args[] = {tileSize};
+    uint64_t args[] = {(uint64_t)tileSize};
     artsGuid_t    mulGpuGuid = artsEdtCreateGpu(mmKernel, artsGetCurrentNode(), 1, args, 3, grid, threads, toSignal, k, cTileGuid);
     artsSignalEdt(mulGpuGuid, 0, aTileGuid);
     artsSignalEdt(mulGpuGuid, 1, bTileGuid);
@@ -302,21 +302,27 @@ void initPerWorker(unsigned int nodeId, unsigned int workerId, int argc, char** 
             if((i * numBlocks + j) % totalThreads == globalThreadId)
             {
 #if GPUMM
-                uint64_t sumArgs[] = {tileSize};
-                dim3 threads (SMTILE, SMTILE);
-                dim3 grid((tileSize+SMTILE-1)/SMTILE, (tileSize+SMTILE-1)/SMTILE);
+              uint64_t sumArgs[] = {(uint64_t)tileSize};
+              dim3 threads(SMTILE, SMTILE);
+              dim3 grid((tileSize + SMTILE - 1) / SMTILE,
+                        (tileSize + SMTILE - 1) / SMTILE);
 
-                artsGuid_t sumGuid = artsEdtCreateGpuPT (sumMMKernel, nodeId, 1, sumArgs, numBlocks, grid, threads, doneGuid, 3 + (i * numBlocks + j), 0);
+              artsGuid_t sumGuid = artsEdtCreateGpuPT(
+                  sumMMKernel, nodeId, 1, sumArgs, numBlocks, grid, threads,
+                  doneGuid, 3 + (i * numBlocks + j), 0);
 #else
                 uint64_t sumArgs[] = {doneGuid, i, j};
                 artsGuid_t sumGuid = artsEdtCreate(sumMM, nodeId, 3, sumArgs, numBlocks);
 #endif
                 for(unsigned int k=0; k<numBlocks; k++)
                 {
-                    uint64_t args[] = {sumGuid, i, j, k};
-                    artsGuid_t mulGuid = artsEdtCreate(multiplyMM, nodeId, 4, args, 2);
-                    artsSignalEdt(mulGuid, 0, artsGetGuid(aTileGuids, i * numBlocks + k));
-                    artsSignalEdt(mulGuid, 1, artsGetGuid(bTileGuids, k * numBlocks + j));
+                  uint64_t args[] = {(uint64_t)sumGuid, i, j, k};
+                  artsGuid_t mulGuid =
+                      artsEdtCreate(multiplyMM, nodeId, 4, args, 2);
+                  artsSignalEdt(mulGuid, 0,
+                                artsGetGuid(aTileGuids, i * numBlocks + k));
+                  artsSignalEdt(mulGuid, 1,
+                                artsGetGuid(bTileGuids, k * numBlocks + j));
                 }
             }
         }
