+++ example/gpu/mm/mmTileNaive.cu
@@ -130,7 +130,7 @@ void multiplyMM(uint32_t paramc, uint64_t * paramv, uint32_t depc, artsEdtDep_t 
     dim3 threads(SMTILE, SMTILE);
     dim3 grid((tile_size+SMTILE-1)/SMTILE, (tile_size+SMTILE-1)/SMTILE);
     
-    uint64_t args[] = {tile_size};
+    uint64_t args[] = {(uint64_t)tile_size};
     artsGuid_t    mulGpuGuid = artsEdtCreateGpu(mmKernel, artsGetCurrentNode(), 1, args, 3, grid, threads, toSignal, k, cTileGuid);
     artsSignalEdt(mulGpuGuid, 0, aTileGuid);
     artsSignalEdt(mulGpuGuid, 1, bTileGuid);
@@ -276,18 +276,22 @@ void initPerWorker(unsigned int nodeId, unsigned int workerId, int argc, char** 
             if((i * numBlocks + j) % totalThreads == globalThreadId)
             {
 #if GPUMM
-                uint64_t sumArgs[] = {tile_size};
-                dim3 threads (tile_size, tile_size);
-                dim3 grid((tile_size+SMTILE-1)/SMTILE, (tile_size+SMTILE-1)/SMTILE);
-                artsGuid_t sumGuid = artsEdtCreateGpuPT (sumMMKernel, nodeId, 1, sumArgs, numBlocks, grid, threads, doneGuid, 3 + (i * numBlocks + j), 0);
+              uint64_t sumArgs[] = {(uint64_t)tile_size};
+              dim3 threads(tile_size, tile_size);
+              dim3 grid((tile_size + SMTILE - 1) / SMTILE,
+                        (tile_size + SMTILE - 1) / SMTILE);
+              artsGuid_t sumGuid = artsEdtCreateGpuPT(
+                  sumMMKernel, nodeId, 1, sumArgs, numBlocks, grid, threads,
+                  doneGuid, 3 + (i * numBlocks + j), 0);
 #else
                 uint64_t sumArgs[] = {doneGuid, i, j};
                 artsGuid_t sumGuid = artsEdtCreate(sumMM, nodeId, 3, sumArgs, numBlocks);
 #endif
                 for(unsigned int k=0; k<numBlocks; k++)
                 {
-                    uint64_t args[] = {sumGuid, i, j, k};
-                    artsGuid_t mulGuid = artsEdtCreate(multiplyMM, nodeId, 4, args, 2);
+                  uint64_t args[] = {(uint64_t)sumGuid, i, j, k};
+                  artsGuid_t mulGuid =
+                      artsEdtCreate(multiplyMM, nodeId, 4, args, 2);
                     artsSignalEdt(mulGuid, 0, aMatGuid);
                     artsSignalEdt(mulGuid, 1, bMatGuid);
                 }
